save tags in parse database outside of fragment so that context is never null
